requirements.txt
numpy
pandas
scikit-learn
requests
spotipy
python-dotenv
